@@ -1,69 +1,70 @@
 {
-  "name": "ngx-card-deck-starter",
-  "author": "mattmutt",
-  "version": "1.0.0",
-  "description": "basic sample starter kit using the ngx-card-deck component",
-  "private": false,
-  "license": "MIT",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mattmutt/ngx-card-deck-starter.git"
-  },
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve --no-aot --configuration=local --source-map=false",
-    "start:prod": "ng serve --prod --aot --configuration=production --build-optimizer=false --sourcemap=true --output-hashing=none",
-    "build": "npm run build:local",
-    "build:local": "ng build --configuration=local --source-map=true",
-    "build:prod": "ng build --prod --base-href=/ngx-card-deck-starter/ --configuration=production --aot --build-optimizer=true --common-chunk=false --source-map=false --output-hashing=none",
-    "lint": "ng lint",
-    "e2e": "ng e2e",
-    "test": "ng test --source-map=false"
-  },
-  "dependencies": {
-    "@angular/animations": "7.2.14",
-    "@angular/common": "7.2.14",
-    "@angular/compiler": "7.2.14",
-    "@angular/core": "7.2.14",
-    "@angular/forms": "7.2.14",
-    "@angular/platform-browser": "7.2.14",
-    "@angular/platform-browser-dynamic": "7.2.14",
-    "@angular/router": "7.2.14",
-    "@clr/angular": "1.1.3",
-    "@clr/icons": "1.1.3",
-    "@clr/ui": "1.1.3",
-    "angular-schema-based-json-editor": "^7.22.0",
-    "ngx-card-deck": "0.0.0-beta.4",
-    "@webcomponents/custom-elements": "^1.2.4",
-    "core-js": "^2.6.5",
-    "resize-observer-polyfill": "^1.5.1",
-    "rxjs": "~6.4.0",
-    "zone.js": "^0.8.29"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.14.0-beta.5",
-    "@angular/cli": "7.2.4",
-    "@angular/compiler-cli": "7.2.14",
-    "@angular/language-service": "7.2.14",
-    "@types/jasmine": "~3.3.9",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~6.0.60",
-    "codelyzer": "^4.5.0",
-    "jasmine-core": "~3.3.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "4.0.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.4",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "node-sass": "4.11.0",
-    "protractor": "~5.4.2",
-    "rxjs-tslint-rules": "^4.23.1",
-    "ts-node": "~8.0.3",
-    "tsdocs": "^1.0.2",
-    "tsickle": "0.34.0",
-    "tslint": "~5.13.1",
-    "typescript": "3.2"
-  }
+    "name": "ngx-card-deck-starter",
+    "author": "mattmutt",
+    "version": "8.0.0",
+    "description": "basic sample starter kit using the ngx-card-deck component",
+    "private": false,
+    "license": "MIT",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/mattmutt/ngx-card-deck-starter.git"
+    },
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve --no-aot --configuration=local --source-map=false",
+        "start:debug": "ng serve --no-aot --configuration=local",
+        "start:prod": "ng serve --prod --aot --configuration=production",
+        "build": "npm run build:local",
+        "build:local": "ng build --configuration=local --source-map=true",
+        "build:prod": "ng build --prod --base-href=/ngx-card-deck-seed/ --configuration=production --aot --build-optimizer=true --common-chunk=false --source-map=false --output-hashing=none",
+        "build:prod:local": "ng build --prod --base-href=/ngx-card-deck-seed/ --configuration=production --aot --preserve-symlinks --build-optimizer=true --common-chunk=false --source-map=false --output-hashing=none",
+        "lint": "ng lint",
+        "e2e": "ng e2e",
+        "test": "ng test"
+    },
+    "dependencies": {
+        "@angular/animations": "^8.1.0",
+        "@angular/common": "^8.1.0",
+        "@angular/compiler": "^8.1.0",
+        "@angular/core": "^8.1.0",
+        "@angular/forms": "^8.1.0",
+        "@angular/platform-browser": "^8.1.0",
+        "@angular/platform-browser-dynamic": "^8.1.0",
+        "@angular/router": "^8.1.0",
+        "@clr/angular": "2.1.0",
+        "@clr/icons": "2.1.0",
+        "@clr/ui": "2.1.0",
+        "@webcomponents/custom-elements": "^1.2.4",
+        "angular-schema-based-json-editor": "^7.23.1",
+        "ngx-card-deck": "^8.0.0",
+        "core-js": "^2.6.9",
+        "resize-observer-polyfill": "^1.5.1",
+        "rxjs": "~6.4.0",
+        "zone.js": "^0.9.1"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "~0.801.2",
+        "@angular/cli": "^8.1.0",
+        "@angular/compiler-cli": "^8.1.0",
+        "@angular/language-service": "^8.1.0",
+        "@types/jasmine": "~3.3.13",
+        "@types/jasminewd2": "~2.0.6",
+        "@types/node": "~10.14.10",
+        "codelyzer": "^5.1.0",
+        "jasmine-core": "~3.3.0",
+        "jasmine-spec-reporter": "~4.2.1",
+        "karma": "4.0.0",
+        "karma-chrome-launcher": "~2.2.0",
+        "karma-cli": "~2.0.0",
+        "karma-coverage-istanbul-reporter": "^2.0.4",
+        "karma-jasmine": "~2.0.1",
+        "karma-jasmine-html-reporter": "^1.4.0",
+        "protractor": "~5.4.2",
+        "rxjs-tslint-rules": "^4.24.3",
+        "ts-node": "~8.3.0",
+        "tsdocs": "^1.0.2",
+        "tsickle": "0.35.0",
+        "tslint": "~5.18.0",
+        "typescript": "3.4"
+    }
 }
